refactor: serviceのディレクトリを削除
