+++ src/app/layout.tsx
@@ -3,7 +3,7 @@ import { ReactQueryDevtools } from '@tanstack/react-query-devtools'
 import type { Metadata } from 'next'
 import { Inter } from 'next/font/google'
 import { SessionProvider } from 'next-auth/react'
-import { QueryClientProviders } from '@/components/TanstackQueryClient/QueryClientProviders'
+import { QueryClientProviders } from '@/components/QueryClientProviders'
 const inter = Inter({ subsets: ['latin'] })
 
 export const metadata: Metadata = {
